requirements.txt
torch
torchvision
gradio
PyYAML
Pillow
numpy
python-telegram-bot==20.7
